+++ src/core/project/core_port.h
@@ -125,6 +125,15 @@ public: // statistics
     inline CoreStatistics statisticsCore() const { QReadLocker locker(&m_statLock); return *m_stat; }
     virtual void resetStatistics();
 
+    inline quint32 statTimeResponseLast () const { QReadLocker locker(&m_statLock); return m_stat->timeResponseLast ; }
+    inline quint32 statTimeResponseMin  () const { QReadLocker locker(&m_statLock); return m_stat->timeResponseMin  ; }
+    inline quint32 statTimeResponseMax  () const { QReadLocker locker(&m_statLock); return m_stat->timeResponseMax  ; }
+    inline quint32 statTimeResponseAvg  () const { QReadLocker locker(&m_statLock); return m_stat->timeResponseAvg  ; }
+    inline quint64 statTimeResponseTotal() const { QReadLocker locker(&m_statLock); return m_stat->timeResponseTotal; }
+    inline quint32 statTimeResponseCount() const { QReadLocker locker(&m_statLock); return m_stat->timeResponseCount; }
+
+    inline void updateResponseTime(quint64 responseTime) { QWriteLocker locker(&m_statLock); m_stat->updateResponseTime(responseTime); }
+
     inline quint32 statCountTx() const { QReadLocker locker(&m_statLock); return m_stat->countTx; }
     inline quint32 statCountRx() const { QReadLocker locker(&m_statLock); return m_stat->countRx; }
     inline quint32 statCountGood() const { QReadLocker locker(&m_statLock); return m_stat->countGood; }