+++ changelog.md
@@ -220,7 +220,7 @@ List of changes of new v0.4 version of ModbusTools:
 * Updated `Send Message` dialog for Modbus client (added predefined message list)
 * Updated `Scanner` dialog for Modbus client (added function support scanning)
 * Added `LogView` message category colorization
-* Added statistics for Port/Device of `mbclient`/`mbserver`
+* Added statistics for Port/Device of `mbclient`/`mbserver`h
 * Renamed `ByteOrder` param into `SwapBytes`
 * Renamed project file extensions `pjc`/`pjs` into `mbc`/`mbs`
 * Renamed `core`/`client`/`server` into `mbcore`/`mbclient`/`mbserver`
@@ -231,4 +231,5 @@ List of changes of new v0.4 version of ModbusTools:
 * Added device settings for write function of single coil/register
 * Added enable/disable for mbclient device (menu `Device/Enable`, context menu)
 * Fixed issue #70: Changing RegisterOrder doesn't affect the value
-* Fixed issue for client's device dialog, when Port type combobox was empty.
+* Fixed issue for client's device dialog, when Port type combobox was empty
+* Added `Response Time` statistics for client